// src/types/itemdefinitions_master_crafting.ts
export type ItemdefsMasterCraftingDatasheet = [{
    itemID: string
    bindOnPickup: boolean
    bindOnEquip: boolean
    minGearScore: number
    maxGearScore: number
    ingredientCategories: string
    ingredientGearScoreBaseBonus: number
    ingredientGearScoreMaxBonus: number
    name: string
}]